+++ Backend/routes/Reverb-Scoopers/index.js
@@ -3,9 +3,8 @@
 const { Router } = require('express');
 const productsIndexRouter = Router();
 
+productsIndexRouter.use('/watchlist', require('./getWatchlist'));
 // productsIndexRouter.use('/listings', require('./getListings'));
 // productsIndexRouter.use('/prices', require('./getPrices'));
-productsIndexRouter.use('/watchlist', require('./getWatchlist'));
-
 
 module.exports = productsIndexRouter;
