+++ Backend/routes/index.js
@@ -3,6 +3,6 @@
 const { Router } = require('express');
 const router = Router();
 
-router.use(require('../Reverb-Scoopers/index.js'));
+router.use(require('./Reverb-Scoopers/index.js'));
 
 module.exports = router;
